move course-related use case tests to isolated test root
