+++ test/java/use_case/CourseUseCases/LeaveCourseUseCase/LeaveCourseUseCaseTest.java
@@ -1,4 +1,4 @@
-package use_case.CourseUseCases.JoinCourseUseCase;
+package java.use_case.CourseUseCases.LeaveCourseUseCase;
 
 import data_access.CourseDataAccessObject;
 import data_access.FileCacheUserDataAccessObject;
@@ -6,17 +6,22 @@ import entity.Course;
 import entity.User;
 import interface_adapter.presenter.CoursePromptPresenter;
 import interface_adapter.viewmodel.CoursePromptViewModel;
+import org.junit.Assert;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import repositories.CourseRepository;
 import repositories.UserRepository;
+import use_case.CourseUseCases.LeaveCourseUseCase.LeaveCourseInputBoundary;
+import use_case.CourseUseCases.LeaveCourseUseCase.LeaveCourseInputData;
+import use_case.CourseUseCases.LeaveCourseUseCase.LeaveCourseOutputBoundary;
+import use_case.CourseUseCases.LeaveCourseUseCase.LeaveCourseUseCase;
 
 import java.io.File;
 import java.io.IOException;
 
-public class JoinCourseUseCaseTest {
+public class LeaveCourseUseCaseTest {
     private String userpath = "test_userCache.json";
     private String coursepath = "test_courses.json";
     private UserRepository userDataAccessObject;
@@ -35,39 +40,11 @@ public class JoinCourseUseCaseTest {
     }
 
     @Test
-    public void testValidJoin() {
+    public void testLeaveCourseValid() {
         Assertions.assertDoesNotThrow(() -> {
             CoursePromptViewModel viewModel = new CoursePromptViewModel();
-            JoinCourseOutputBoundary presenter = new CoursePromptPresenter(null, viewModel, null);
-            JoinCourseInputBoundary interactor = new JoinCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
-
-            User user = new User("Test User", "TestPassword1", new User[]{}, new Course[]{});
-            Course course = new Course("Test Course", "Test Description");
-
-            userDataAccessObject.WriteToCache(user);
-            courseDataAccessObject.WriteToCache(course);
-
-            Assertions.assertFalse(course.containsUser(user.getUsername()));
-
-            JoinCourseInputData inputData = new JoinCourseInputData(
-                    user.getUsername(),
-                    course.getName()
-            );
-
-            interactor.execute(inputData);
-
-            Assertions.assertTrue(viewModel.getResponse());
-            Assertions.assertFalse(userDataAccessObject.ReadFromCache("Test User").getCourses().isEmpty());
-            Assertions.assertTrue(courseDataAccessObject.findByName("Test Course").containsUser(user.getUsername()));
-        });
-    }
-
-    @Test
-    public void testAlreadyInCourse() {
-        Assertions.assertDoesNotThrow(() -> {
-            CoursePromptViewModel viewModel = new CoursePromptViewModel();
-            JoinCourseOutputBoundary presenter = new CoursePromptPresenter(null, viewModel, null);
-            JoinCourseInputBoundary interactor = new JoinCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
+            LeaveCourseOutputBoundary presenter = new CoursePromptPresenter(null, null, viewModel);
+            LeaveCourseInputBoundary interactor = new LeaveCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
 
             User user = new User("Test User", "TestPassword1", new User[]{}, new Course[]{});
             Course course = new Course("Test Course", "Test Description");
@@ -75,12 +52,42 @@ public class JoinCourseUseCaseTest {
             user.addCourse(course);
             course.addUser(user);
 
-            Assertions.assertTrue(course.containsUser(user.getUsername()));
+            userDataAccessObject.WriteToCache(user);
+            courseDataAccessObject.WriteToCache(course);
+
+            Assertions.assertTrue(course.containsUser("Test User"));
+            Assertions.assertTrue(user.getCourses().contains("Test Course"));
+
+            LeaveCourseInputData inputData = new LeaveCourseInputData(
+                    user.getUsername(),
+                    course.getName()
+            );
+
+            interactor.execute(inputData);
+
+            Assertions.assertTrue(viewModel.getResponse());
+            Assertions.assertTrue(userDataAccessObject.ReadFromCache("Test User").getCourses().isEmpty());
+            Assertions.assertEquals(0, courseDataAccessObject.findByName("Test Course").getUserNames().length);
+        });
+    }
+
+    @Test
+    public void testNotInCourse() {
+        Assertions.assertDoesNotThrow(() -> {
+            CoursePromptViewModel viewModel = new CoursePromptViewModel();
+            LeaveCourseOutputBoundary presenter = new CoursePromptPresenter(null, null, viewModel);
+            LeaveCourseInputBoundary interactor = new LeaveCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
+
+            User user = new User("Test User", "TestPassword1", new User[]{}, new Course[]{});
+            Course course = new Course("Test Course", "Test Description");
 
             userDataAccessObject.WriteToCache(user);
             courseDataAccessObject.WriteToCache(course);
 
-            JoinCourseInputData inputData = new JoinCourseInputData(
+            Assertions.assertFalse(course.containsUser("Test User"));
+            Assertions.assertFalse(user.getCourses().contains("Test Course"));
+
+            LeaveCourseInputData inputData = new LeaveCourseInputData(
                     user.getUsername(),
                     course.getName()
             );
@@ -88,6 +95,8 @@ public class JoinCourseUseCaseTest {
             interactor.execute(inputData);
 
             Assertions.assertFalse(viewModel.getResponse());
+            Assertions.assertTrue(userDataAccessObject.ReadFromCache("Test User").getCourses().isEmpty());
+            Assertions.assertEquals(0, courseDataAccessObject.findByName("Test Course").getUserNames().length);
         });
     }
 
@@ -95,14 +104,16 @@ public class JoinCourseUseCaseTest {
     public void testCourseDoesNotExist() {
         Assertions.assertDoesNotThrow(() -> {
             CoursePromptViewModel viewModel = new CoursePromptViewModel();
-            JoinCourseOutputBoundary presenter = new CoursePromptPresenter(null, viewModel, null);
-            JoinCourseInputBoundary interactor = new JoinCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
+            LeaveCourseOutputBoundary presenter = new CoursePromptPresenter(null, null, viewModel);
+            LeaveCourseInputBoundary interactor = new LeaveCourseUseCase(presenter, userDataAccessObject, courseDataAccessObject);
 
             User user = new User("Test User", "TestPassword1", new User[]{}, new Course[]{});
 
             userDataAccessObject.WriteToCache(user);
 
-            JoinCourseInputData inputData = new JoinCourseInputData(
+            Assertions.assertEquals(0, user.getCourses().size());
+
+            LeaveCourseInputData inputData = new LeaveCourseInputData(
                     user.getUsername(),
                     "Test Course"
             );
